Add files via upload
This is the first step in a work in progress to almost fully automating the process of data submission. It currently only works with iOS screenshots of PokeGenie, but I have been working on an Android version and trying to learn OpenCV to perfect it even more because it currently still has quite a few bugs that make make me skeptical, but with just brute-force testing and a lot of frustration, I've been slowly improving it.